Move WebView tracking into a separated feature package
