+++ dd-sdk-android/apiSurface
@@ -1213,9 +1213,6 @@ interface com.datadog.android.rum.tracking.TrackingStrategy
 interface com.datadog.android.rum.tracking.ViewAttributesProvider
   fun extractAttributes(android.view.View, MutableMap<String, Any?>)
 interface com.datadog.android.rum.tracking.ViewTrackingStrategy : TrackingStrategy
-class com.datadog.android.rum.webview.DatadogEventBridge
-  fun send(String)
-  fun getAllowedWebViewHosts(): String
 open class com.datadog.android.rum.webview.RumWebChromeClient : android.webkit.WebChromeClient
   constructor()
   override fun onConsoleMessage(android.webkit.ConsoleMessage?): Boolean
@@ -1302,4 +1299,7 @@ data class com.datadog.android.tracing.model.SpanEvent
     fun toJson(): com.google.gson.JsonElement
     companion object 
       fun fromJson(kotlin.String): SimCarrier
+class com.datadog.android.webview.DatadogEventBridge
+  fun send(String)
+  fun getAllowedWebViewHosts(): String
 class com.datadog.tools.annotation.NoOpImplementation
