+++ dd-sdk-android/src/main/kotlin/com/datadog/android/webview/internal/WebRumEventMapper.kt
@@ -4,7 +4,7 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.android.rum.webview
+package com.datadog.android.webview.internal
 
 import com.datadog.android.core.internal.CoreFeature
 import com.datadog.android.rum.internal.domain.RumContext
