+++ dd-sdk-android/src/test/kotlin/com/datadog/android/webview/internal/WebLogEventConsumerTest.kt
@@ -4,7 +4,7 @@
  * Copyright 2016-Present Datadog, Inc.
  */
 
-package com.datadog.android.rum.internal.webview
+package com.datadog.android.webview.internal
 
 import android.util.Log
 import com.datadog.android.core.internal.CoreFeature
@@ -13,8 +13,6 @@ import com.datadog.android.log.LogAttributes
 import com.datadog.android.log.internal.logger.LogHandler
 import com.datadog.android.log.model.LogEvent
 import com.datadog.android.rum.internal.domain.RumContext
-import com.datadog.android.rum.webview.WebLogEventConsumer
-import com.datadog.android.rum.webview.WebRumEventContextProvider
 import com.datadog.android.utils.assertj.DeserializedLogEventAssert
 import com.datadog.android.utils.forge.Configurator
 import com.datadog.android.utils.mockSdkLogHandler
